Added SongInformationView
Added SongInformationView and DurationView
